mmc: omap: Typecast pdata->max_freq
max_freq passed as part of mmc platform data
struct omap2_hsmmc_info.max_freq is "unsigned long"
where as struct mmc_host.f_max is "unsigned int".

So, type casting to remove compile warnings.

+++ drivers/mmc/host/omap.c
@@ -1330,7 +1330,8 @@ static int __init mmc_omap_new_slot(struct mmc_omap_host *host, int id)
 	else
 		mmc->f_max = 24000000;
 	if (host->pdata->max_freq)
-		mmc->f_max = min(host->pdata->max_freq, mmc->f_max);
+		mmc->f_max = min((unsigned int)host->pdata->max_freq,
+							mmc->f_max);
 	mmc->ocr_avail = slot->pdata->ocr_mask;
 
 	/* Use scatterlist DMA to reduce per-transfer costs.
